Refactor matrix into new module
Fix up a couple test bugs and do some other refactor / cleanup
